Split J2KECorrection.h off J2OrbitalSoA.h
